Remove partial functions from reexports
This removes partial and unsafe functions from reexport modules, and
moves them to their own modules.  As a side-effect of this,
documentation in many modules has been improved as there are now
explicit export lists.

I have not touched RIO.Process as I am entirely unfamiliar with the API.
To the best of my knowledge, RIO.Directory, RIO.FilePath, and RIO.Time
do not contain any partial functions so are also not touched.

There are various CPP conditions in the export lists for Set, Map,
and the Vector modules (possibly among others) for compatibility with
lts-7.  It may instead be preferable to either drop the newer functions
or drop some backwards compatibility to avoid conditional reexports.

Resolves #43