Add logGeneric function that receives CallStack
Sometimes we want to perform logging in a context and not
emit to the output handle directly, but instead, we want to
deliver it to a thread that performs the actual logging.

Without this function, we will get the CallStack of the logging
thread, rather than where it was invoked.