Fix #205 On Windows, findExecutable uses ProcessContext consistently
Also, refactor `findExecutable`; using new helper functions `addPcExeExtensions` and `testFPs` to eliminate duplication in code, and applying `hlint` suggestion.

Also, on POSIX, the `System.Directory.executable` test is now applied consistently to both limbs of `findExecutable`.

On Windows, the `findExecutable name =` limb of `findExecutable` no longer filters by the `System.Directory.executable` test and, so, takes into account fully the list of executable extensions in the `ProcessContext`.

This also extends the Haddock documentation for `mkDefaultProcessContext` (which expressly now refers back to 'mkProcessContext') and `findExecutable`, and corrects the Haddock documentation for `modifyEnvVars`.

This extends the (hidden) documentation for `mkProcessContext`, explaining the `defaultPATHEXT` used previously.

On Windows, an excutable with a filename with no extension cannot be executed from the command line. `mkProcessContext' and `findExecutable` are modified accordingly. `mkProcessContext` now excludes `""` from the head of the list of extensions and the limbs of `findExecutable` use a `isWindows && FP.hasExtension` test to decide whether to include the name itself in the list of names tried.

This also adds, and exports from the module, `exeExecutables`, following the pattern of the Haddock documentation for `System.Directory.exeExecutable` (with logical amendments).

The `isWindows` function is moved to be a generally-available helper function.